Optimize final VHD
This adds a -ReadOnly switch to MountVHDandRunBlock

After the final VHD is created, it will be mounted read only, and
Optimize-VHD run against it twice.

Mounting the VHD read only is required for some of the optimizations
Optimize-VHD does.

Why do I do it twice? I have no idea. But extensive testing on my Windows 8.1 host
system shows that doing it once does nothing, but twice shrinks the file
substantially.

This saves 100-200MB in each image.