Optionally clean WinSXS store
Use dism.exe to clean the component store of the image. This can save
another couple hundred MB of space, depending on how many updates there
are.  Disabled by default because it takes a long time to run - about
20-30 minutes per image on my test system.

Enabling it will remove any superceded components entirely - updates will
not be able to be uninstalled anymore.