Allow dot sourcing the script
This allows the script to be dot sourced, so it can be executed as part
of another script.  The biggest reason for this change is so I can set
my own parameters of which images to build without editing Factory.ps1 -
making updating easier.

I changed the name of the main function to Start-ImageFactory to follow
the standard powershell naming convention, and added comment based help.
This isn't strictly necessary, but since the function is now intended to
be usable by other scripts, it makes sense to me to follow the Verb-Noun
naming standard.

By default this changes nothing - the script runs the same way it always
did if $StartFactory = $true.  When $StartFactory = $false, it won't run
anything automatically and can be dot sourced in an outside script
instead.