Update manage_images.md
Given that we currently only manage container images in either powershell OR docker, when running a powershell cmdlet to install a BASEOS image, naturally one would assume there is a DOCKER command to do the same.  However BASEOS images are available to both management tool sets.  it is not clear here.  What further complicates the issue is that when an image is added, it doesn't immediately show up in DOCKER, and only showed up after I had restarted the Daemon.  Probably known issue, probably not going to be broken in GA.  Definitely useful now, would have saved me 30 / 40 minutes of searching.