Typo fix (#841)
* Typo fix

Typo fix. It is called "Windows 10 Pro", not "Windows 10 Professional".

* Typo fix

Typo fix. It is called "Windows 10 Pro", not "Windows 10 Professional". Also, Windows 8 did not have an "Education" edition or "Home" edition. (The equivalent to "Home" edition in Windows 8 had not name.) Finally, this page is supposed to be exclusively about Windows 10.

* Typo fix

Typo fix. It is called "Windows 10 Pro", not "Windows 10 Professional".

* Typo fix

Typo fix. It is called "Windows 10 Pro", not "Windows 10 Professional". Windows 8 does not have an Education edition. And this page is about Windows 10 only.