"Get-HNSNetwork" cmdlet never works on WS2016
On section subheading "Clean up Docker data and system components",  "Get-HNSNetwork | Remove-HNSNetwork" cmdlet works fine on Windows 10, but it never works on Windows Server 2016. You should use  "Get-ContainerNetwork | Remove-ContainerNetwork" in place of "Get-HNSNetwork | Remove-HNSNetwork" on Windows Server 2016. My customer faced this issue on Windows Server 2016 and complained that this article was not correct.