Showing Location for Already-Assigned Devices
The script as is doesn't display the path for already-assigned devices. This is something you need to know if you already assigned a device over a year ago and have no memory of this place :)

Simple fix just moves the $locationpath variable up the chain and prints the path immediately after the friendly name regardless. Also removes the $locationpath prints later in the script since it has this new upfront print. Changes text to have black background and yellow foreground.