Defined Actionable Instruction
This article invents a term "Actionable Instruction". This term is found only in this article or in blog posts the plagiarized sentences from this post. I defined what the actionable instructions are RUN, COPY, ADD -- the instructions that create a new layer when invoked. All other instructions create temporary intermediate images and therefor do not increase the size of the build.