Update SurveyDDA.ps1
The cmdlet "Get-WmiObject" with, used in "SurveyDDA.ps1" as the alias "gwmi," was superseded by "Get-CimInstance" in Powershell 3.0 and later.  This causes calls the calls to "gwmi" the occur in the script to fail starting with Powershell 3.0. This update replaces "gwmi" with "Get-CimInstance" for compatibility with these later Powershell versions.